fix: update vulnerable modules (#430)
I ran an `npm audit fix` on this and several packages were updated. The
ones that can't be appear to be dependent on `speccy` which may not be
able to be updated. That's a dev dependency so I'm not sure that it's a
critical situation in our use case.

I ran the tests afterward and they all passed.

JIRA: RHCLOUD-11294
BZ: 1907860

Signed-off-by: Stephen Adams <tsadams@gmail.com>